Universal/RequireFinalMethodsInTraits: move sniff to `FunctionDeclarations` category
... which feels more appropriate.
