+++ Universal/Tests/FunctionDeclarations/RequireFinalMethodsInTraitsUnitTest.inc.fixed
@@ -96,7 +96,7 @@ trait FixMe {
     /*comment*/ protected final function withCommentBeforeKeyword() {}
 }
 
-// phpcs:set Universal.OOStructures.RequireFinalMethodsInTraits includeMagicMethods true
+// phpcs:set Universal.FunctionDeclarations.RequireFinalMethodsInTraits includeMagicMethods true
 trait MagicMethodsAreFlaggedOnRequest {
     public final function __construct() {}
     public final function __destruct() {}
@@ -117,4 +117,4 @@ trait MagicMethodsAreFlaggedOnRequest {
 }
 
 // Reset property to default value.
-// phpcs:set Universal.OOStructures.RequireFinalMethodsInTraits includeMagicMethods false
+// phpcs:set Universal.FunctionDeclarations.RequireFinalMethodsInTraits includeMagicMethods false
